Fix baseurl and permalink relative path issues
Replaces "remove: '/'" to "remove_first: '/'" when preceded by
site.baseurl.

Attempting to build a site with permalinks within a subdirectory
generates incorrect links. Links with a leading slash were being
interpreted as a complete link with a relative schema because another
slash was being added to the URL.  Slashes in links are missing when
using permalink schemes other than /:title because all '/' characters
were being removed from the pathname.

To alleviate this, we only remove the first '/' to keep browsers from
misinterpreting the path as an absolute URL. Now, all permalink schemes
work as intended.